feat: shared settings lib, LauncherService, hub/docs
Add DeckTop.Settings for unified settings.toml; extract ILauncherService; improve BackToWindows cleanup; hub MKB + READMEs; deferred init and optional GitHub update check; refresh AGENTS and HANDOFF.

Made-with: Cursor
