+++ gcmloader/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using AppSettings = DeckTop.Settings.AppSettings;
