+++ gcmloader/Services/FocusNavigationHelper.cs
@@ -1,0 +1,17 @@
+using Microsoft.UI.Xaml.Input;
+using Windows.System;
+
+namespace gcmloader.Services;
+
+/// <summary>
+/// Optional shared helpers for directional / Escape keyboard routing (hub ↔ gcmloader parity).
+/// Extend when consolidating custom FocusArea-style logic.
+/// </summary>
+public static class FocusNavigationHelper
+{
+    /// <summary>Returns true if the key is Escape (hardware or Gamepad B).</summary>
+    public static bool IsCancelKey(KeyRoutedEventArgs e)
+    {
+        return e.Key == VirtualKey.Escape;
+    }
+}
