+++ gcmloader/gcmloaderwindow.xaml.cs
@@ -61,9 +61,12 @@ using Task = System.Threading.Tasks.Task;
 
 namespace gcmloader
 {
+    // Heavy secondary UI (“islands”) should load on demand (Loaded / navigate), not in the shell ctor.
 
     public sealed partial class MainWindow : Window
     {
+        private readonly ILauncherService _launcherService;
+
         #region needed
         private void BoostProcessPriority()
         {
@@ -2367,6 +2370,17 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
            // Logger.Initialize();
             this.InitializeComponent();
 
+            _launcherService = new LauncherService(new LauncherShell
+            {
+                MakeSelfNonTopmost = MakeSelfNonTopmost,
+                GetHostWindowHandle = () => WinRT.Interop.WindowNative.GetWindowHandle(this),
+                FindSteamBigPictureWindow = FindSteamBigPictureWindow,
+                ForcefullyBringToForeground = ForcefullyBringToForeground,
+                ShowMessageAsync = async msg => await messagebox(msg),
+                BackToWindows = BackToWindows,
+                StartXboxAsync = StartXbox,
+            });
+
             _keyboardNavigationEnabled = !TryLoadControllerNavigationEnabled();
             _suppressControllerNavigationToggle = true;
             if (ControllerNavigationToggle != null)
@@ -4601,10 +4615,40 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
         }
         private void BackToWindows()
         {
-            // TODO: restore taskbar, re-enable services, cleanup audio
-            // Shell registration and UAC steps removed — DeckTop is not a shell replacement
-            Logger.Log("BackToWindows called — graceful exit");
-            Environment.Exit(0);
+            // No Winlogon / shell swap / UAC — DeckTop is not a shell replacement.
+            try
+            {
+                Logger.Log("BackToWindows — cleanup before exit");
+
+                try { _focusCheckTimer?.Stop(); } catch { }
+                try { _minimizeGracePeriodTimer?.Stop(); } catch { }
+                try { _mouseIdleTimer?.Stop(); } catch { }
+                try { _statusUpdateTimer?.Stop(); } catch { }
+                try { _displayWatchdogTimer?.Stop(); } catch { }
+                try { _windowEngineTimer?.Stop(); } catch { }
+                try { _autoMouseTimer?.Stop(); } catch { }
+                try { _taskRefreshTimer?.Stop(); } catch { }
+                try { _timer?.Stop(); } catch { }
+                try { _refreshTimer?.Stop(); } catch { }
+
+                try { TaskbarManager.RestoreOriginalState(); }
+                catch (Exception ex) { Debug.WriteLine($"[Exit] TaskbarManager.RestoreOriginalState: {ex.Message}"); }
+
+                try { TaskbarVisibility.ShowTaskbar(); }
+                catch (Exception ex) { Debug.WriteLine($"[Exit] TaskbarVisibility.ShowTaskbar: {ex.Message}"); }
+
+                try { RenameSteamStartupVideo_End(); }
+                catch (Exception ex) { Debug.WriteLine($"[Exit] RenameSteamStartupVideo_End: {ex.Message}"); }
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine($"[Exit] BackToWindows: {ex.Message}");
+            }
+            finally
+            {
+                try { CleanupLogging(); } catch { }
+                Environment.Exit(0);
+            }
         }
 
         //xbox
@@ -4876,158 +4920,16 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
                 }
             }
         }
-        // Sucht vollautomatisch nach den Pfaden, OHNE die settings.toml zu nutzen!
-        // --- HELPER: Auto-Detect Launcher Paths ---
-        // Automatically finds the executable paths for supported launchers without checking settings
-        // --- HELPER: Auto-Detect Launcher Paths ---
-        public static string AutoDetectLauncherPath(string launcher)
-        {
-            string localAppData = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-            string appData = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
-            string path = null;
+        public static string? AutoDetectLauncherPath(string launcher) => LauncherService.AutoDetectLauncherPath(launcher);
 
-            switch (launcher)
-            {
-                case "steam":
-                    path = RegistryOperations.TryGetSteamExePath();
-                    break;
-
-                case "playnite":
-                    // 1. Zuerst exakt in deinem Standard-Pfad schauen (99% der Fälle)
-                    string defaultPath = Path.Combine(localAppData, "Playnite", "Playnite.FullscreenApp.exe");
-                    if (File.Exists(defaultPath))
-                    {
-                        path = defaultPath;
-                        break;
-                    }
-
-                    // 2. Falls es woanders installiert wurde: Registry checken
-                    string[] playniteRegPaths = {
-                @"SOFTWARE\Playnite",
-                @"SOFTWARE\WOW6432Node\Playnite",
-                @"SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\Playnite",
-                @"SOFTWARE\WOW6432Node\Microsoft\Windows\CurrentVersion\Uninstall\Playnite",
-                @"SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\{14AB2B56-32A1-4F29-BEE2-0BA851CC07C7}_is1" // Playnite Installer GUID
-            };
-
-                    RegistryKey[] roots = { Registry.CurrentUser, Registry.LocalMachine };
-                    foreach (var root in roots)
-                    {
-                        foreach (var regPath in playniteRegPaths)
-                        {
-                            using (var key = root.OpenSubKey(regPath))
-                            {
-                                if (key != null)
-                                {
-                                    string installDir = (key.GetValue("InstallPath") as string) ?? (key.GetValue("InstallLocation") as string);
-                                    if (!string.IsNullOrEmpty(installDir))
-                                    {
-                                        string fullPath = Path.Combine(installDir, "Playnite.FullscreenApp.exe");
-                                        if (File.Exists(fullPath))
-                                        {
-                                            path = fullPath;
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                            if (path != null) break;
-                        }
-                        if (path != null) break;
-                    }
-                    break;
-
-                case "gfn":
-                    path = Path.Combine(localAppData, "NVIDIA Corporation", "GeForceNOW", "CEF", "GeForceNOW.exe");
-                    if (!File.Exists(path))
-                    {
-                        string lnkPath = Path.Combine(appData, @"Microsoft\Windows\Start Menu\Programs\NVIDIA GeForce NOW.lnk");
-                        if (File.Exists(lnkPath)) path = lnkPath; // Der .lnk Trick für GFN funktioniert, da hier keine Fullscreen/Desktop Unterscheidung existiert
-                    }
-                    break;
-            }
-
-            if (path != null && File.Exists(path)) return path;
-
-            return null;
-        }
         private async void SwitchToSpecificLauncher(string launcherId)
         {
-            MakeSelfNonTopmost();
-            Debug.WriteLine($"[GCM] Quick-Launch ausgelöst für: '{launcherId}'...");
-
-            try
-            {
-                switch (launcherId)
-                {
-                    case "steam":
-                        await StartSteam(false); // <--- false: Sanfter Wechsel!
-                        break;
-                    case "gfn":
-                        await StartGfn();
-                        break;
-                    case "xbox":
-                        await StartXbox();
-                        break;
-                    case "playnite":
-                        // FIX: Kein Protokoll-Befehl mehr! Direkt unsere Fullscreen-Methode aufrufen.
-                        await StartPlaynite();
-                        break;
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[GCM] Fehler beim Quick-Launch: {ex.Message}");
-            }
+            await _launcherService.SwitchToSpecificLauncherAsync(launcherId).ConfigureAwait(true);
         }
 
         private async void SwitchToConfiguredLauncher()
         {
-            MakeSelfNonTopmost();
-
-            string launcher = AppSettings.Load<string>("launcher");
-            Debug.WriteLine($"[GCM] Wechsle zu konfiguriertem Launcher: '{launcher}'...");
-
-            try
-            {
-                switch (launcher)
-                {
-                    case "steam":
-                        await StartSteam(false); 
-                        return;
-
-                    case "gfn":
-                        await StartGfn();
-                        return;
-
-                    case "playnite":
-                        // FIX: Auch hier jeden Protokoll-Befehl entfernt!
-                        await StartPlaynite();
-                        return;
-
-                    case "custom":
-                        string customPath = AppSettings.Load<string>("customlauncherpath");
-                        string customProcessName = Path.GetFileNameWithoutExtension(customPath);
-                        Process customProc = Process.GetProcessesByName(customProcessName).FirstOrDefault();
-                        if (customProc != null && customProc.MainWindowHandle != IntPtr.Zero)
-                        {
-                            await ForcefullyBringToForeground(customProc.MainWindowHandle);
-                        }
-                        else
-                        {
-                            await StartOtherLauncher();
-                        }
-                        return;
-
-                    case "xbox":
-                        await StartXbox(false);
-                        return;
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[GCM] Fehler während des Launcher-Wechsels: {ex.Message}");
-            }
+            await _launcherService.SwitchToConfiguredLauncherAsync().ConfigureAwait(true);
         }
         private static void MaximizeXboxWindow(IntPtr hwnd)
         {
@@ -5343,492 +5245,30 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
             switch (launcher)
             {
                 case "steam":
-                    await StartSteam();
+                    await _launcherService.StartSteamAsync(false).ConfigureAwait(false);
                     break;
                 case "playnite":
-                    await StartPlaynite();
+                    await _launcherService.StartPlayniteAsync().ConfigureAwait(false);
                     break;
                 case "custom":
-                    await StartOtherLauncher();
+                    await _launcherService.StartOtherLauncherAsync().ConfigureAwait(false);
                     break;
                 case "gfn":
-                    await StartGfn();
+                    await _launcherService.StartGfnAsync().ConfigureAwait(false);
                     break;
                 case "xbox":
                     await StartXbox(true);
                     break;
                 default:
                     AppSettings.Save("launcher", "steam");
-                    await StartSteam();
+                    await _launcherService.StartSteamAsync(false).ConfigureAwait(false);
                     break;
             }
         }
         #endregion winparts
 
         #endregion functions
-        #region launcher
 
-        #region non-admin launch helper
-        private const uint TOKEN_DUPLICATE_ACCESS = 0x0002;
-        private const uint MAXIMUM_ALLOWED_ACCESS = 0x02000000;
-        private const uint CREATE_UNICODE_ENVIRONMENT_FLAG = 0x00000400;
-
-        private enum SECURITY_IMPERSONATION_LEVEL_ENUM { SecurityAnonymous = 0, SecurityIdentification = 1, SecurityImpersonation = 2, SecurityDelegation = 3 }
-        private enum TOKEN_TYPE_ENUM { TokenPrimary = 1, TokenImpersonation = 2 }
-
-        [StructLayout(LayoutKind.Sequential)]
-        private struct SECURITY_ATTRIBUTES_TOKEN
-        {
-            public int nLength;
-            public IntPtr lpSecurityDescriptor;
-            public bool bInheritHandle;
-        }
-
-        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-        private struct STARTUPINFOW
-        {
-            public int cb;
-            public string lpReserved;
-            public string lpDesktop;
-            public string lpTitle;
-            public uint dwX, dwY, dwXSize, dwYSize;
-            public uint dwXCountChars, dwYCountChars;
-            public uint dwFillAttribute, dwFlags;
-            public short wShowWindow, cbReserved2;
-            public IntPtr lpReserved2;
-            public IntPtr hStdInput, hStdOutput, hStdError;
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        private struct PROCESS_INFORMATION_W
-        {
-            public IntPtr hProcess;
-            public IntPtr hThread;
-            public uint dwProcessId;
-            public uint dwThreadId;
-        }
-
-        [DllImport("advapi32.dll", SetLastError = true)]
-        private static extern bool OpenProcessToken(IntPtr processHandle, uint desiredAccess, out IntPtr tokenHandle);
-
-        [DllImport("wtsapi32.dll", SetLastError = true)]
-        private static extern bool WTSQueryUserToken(uint sessionId, out IntPtr phToken);
-
-        [DllImport("kernel32.dll")]
-        private static extern uint WTSGetActiveConsoleSessionId();
-
-        [DllImport("advapi32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        private static extern bool DuplicateTokenEx(IntPtr hExistingToken, uint dwDesiredAccess, ref SECURITY_ATTRIBUTES_TOKEN lpTokenAttributes, SECURITY_IMPERSONATION_LEVEL_ENUM ImpersonationLevel, TOKEN_TYPE_ENUM TokenType, out IntPtr phNewToken);
-
-        [DllImport("advapi32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
-        private static extern bool CreateProcessWithTokenW(IntPtr hToken, uint dwLogonFlags, string lpApplicationName, string lpCommandLine, uint dwCreationFlags, IntPtr lpEnvironment, string lpCurrentDirectory, ref STARTUPINFOW lpStartupInfo, out PROCESS_INFORMATION_W lpProcessInformation);
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        private static extern bool CloseHandle(IntPtr hObject);
-
-        
-        private static void StartProcessAsNonAdmin(string filePath, string arguments = null, string workingDirectory = null)
-        {
-            IntPtr userToken = GetNonElevatedUserToken();
-
-            if (userToken == IntPtr.Zero)
-            {
-                // Standard launch (will inherit elevated token)
-                Debug.WriteLine("[GCM] StartProcessAsNonAdmin: could not obtain non-elevated token, falling back to elevated launch.");
-                var psi = new ProcessStartInfo(filePath) { UseShellExecute = true };
-                if (!string.IsNullOrEmpty(arguments)) psi.Arguments = arguments;
-                if (!string.IsNullOrEmpty(workingDirectory)) psi.WorkingDirectory = workingDirectory;
-                Process.Start(psi);
-                return;
-            }
-
-            IntPtr duplicateToken = IntPtr.Zero;
-            try
-            {
-                var sa = new SECURITY_ATTRIBUTES_TOKEN { nLength = Marshal.SizeOf<SECURITY_ATTRIBUTES_TOKEN>() };
-
-                if (!DuplicateTokenEx(userToken, MAXIMUM_ALLOWED_ACCESS, ref sa,
-                    SECURITY_IMPERSONATION_LEVEL_ENUM.SecurityImpersonation, TOKEN_TYPE_ENUM.TokenPrimary, out duplicateToken))
-                    throw new InvalidOperationException($"DuplicateTokenEx failed: {Marshal.GetLastWin32Error()}");
-
-                var si = new STARTUPINFOW
-                {
-                    cb = Marshal.SizeOf<STARTUPINFOW>(),
-                    lpDesktop = "winsta0\\default"
-                };
-
-                string cmdLine = string.IsNullOrEmpty(arguments)
-                    ? $"\"{filePath}\""
-                    : $"\"{filePath}\" {arguments}";
-
-                if (!CreateProcessWithTokenW(duplicateToken, 0, null, cmdLine,
-                    CREATE_UNICODE_ENVIRONMENT_FLAG, IntPtr.Zero, workingDirectory, ref si, out var pi))
-                    throw new InvalidOperationException($"CreateProcessWithTokenW failed: {Marshal.GetLastWin32Error()}");
-
-                if (pi.hProcess != IntPtr.Zero) CloseHandle(pi.hProcess);
-                if (pi.hThread != IntPtr.Zero) CloseHandle(pi.hThread);
-            }
-            finally
-            {
-                CloseHandle(userToken);
-                if (duplicateToken != IntPtr.Zero) CloseHandle(duplicateToken);
-            }
-        }
-
-        
-        private static IntPtr GetNonElevatedUserToken()
-        {
-            // 1. Try explorer first 
-            IntPtr token = TryGetTokenFromProcessName("explorer");
-            if (token != IntPtr.Zero) return token;
-
-            // 2. Try other non-elevated user-session processes 
-            string[] candidates = { "RuntimeBroker", "sihost", "ctfmon", "taskhostw", "ShellExperienceHost" };
-            foreach (string name in candidates)
-            {
-                token = TryGetTokenFromProcessName(name);
-                if (token != IntPtr.Zero)
-                {
-                    Debug.WriteLine($"[GCM] GetNonElevatedUserToken: obtained token from '{name}'.");
-                    return token;
-                }
-            }
-
-            // 3. WTSQueryUserToken — the last chance :)
-            try
-            {
-                uint sessionId = WTSGetActiveConsoleSessionId();
-                if (sessionId != 0xFFFFFFFF && WTSQueryUserToken(sessionId, out token))
-                {
-                    Debug.WriteLine($"[GCM] GetNonElevatedUserToken: obtained token via WTSQueryUserToken (session {sessionId}).");
-                    return token;
-                }
-                Debug.WriteLine($"[GCM] WTSQueryUserToken failed: {Marshal.GetLastWin32Error()}");
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[GCM] WTSQueryUserToken exception: {ex.Message}");
-            }
-
-            return IntPtr.Zero;
-        }
-
-        private static IntPtr TryGetTokenFromProcessName(string processName)
-        {
-            try
-            {
-                Process? proc = Process.GetProcessesByName(processName).FirstOrDefault();
-                if (proc == null) return IntPtr.Zero;
-
-                if (OpenProcessToken(proc.Handle, TOKEN_DUPLICATE_ACCESS, out IntPtr token))
-                    return token;
-            }
-            catch { }
-            return IntPtr.Zero;
-        }
-        #endregion non-admin launch helper
-
-        private const uint SWP_NOACTIVATE = 0x0010;
-        // Wir fügen den Parameter 'forceRestart' hinzu. Standard ist 'false'.
-        private async Task StartSteam(bool forceRestart = false)
-        {
-            try
-            {
-                string steamExePath = AutoDetectLauncherPath("steam");
-                if (string.IsNullOrWhiteSpace(steamExePath))
-                    throw new FileNotFoundException("Steam could not be found automatically on this system.");
-
-                // Push GCM to the background to make room for Steam
-                IntPtr myHwnd = WinRT.Interop.WindowNative.GetWindowHandle(this);
-                SetWindowPos(myHwnd, HWND_BOTTOM, 0, 0, 0, 0, SWP_NOMOVE | SWP_NOSIZE | SWP_NOACTIVATE);
-
-                bool steamLiefSchon = false;
-
-                // Check if Steam is already running right at the beginning
-                var steamProc = Process.GetProcessesByName("steam").FirstOrDefault();
-                bool isSteamRunning = steamProc != null;
-
-                // A "Cold Start" is required if explicitly requested (e.g., on boot) OR if Steam is completely closed
-                bool isColdStart = forceRestart || !isSteamRunning;
-
-                // --- DECKY LOADER INTEGRATION ---
-                bool useDeckyLoader = false;
-                try
-                {
-                    useDeckyLoader = AppSettings.Load<bool>("usedeckyloader");
-                }
-                catch
-                {
-                    // Fallback in case the setting doesn't exist yet
-                    Debug.WriteLine("[GCM] 'usedeckyloader' setting not found, defaulting to false.");
-                }
-
-                // We ONLY reset Decky Loader if we are doing a cold start.
-                // If we are just switching via the Launcher Card (Warmstart), we skip this heavy process!
-                if (useDeckyLoader && isColdStart)
-                {
-                    Debug.WriteLine("[GCM] Decky Loader enabled & Cold Boot required. Preparing environment...");
-
-                    // Make sure Steam is completely dead before doing anything with Decky
-                    var allSteamProcs = Process.GetProcessesByName("steam")
-                                               .Concat(Process.GetProcessesByName("steamwebhelper"))
-                                               .ToList();
-
-                    if (allSteamProcs.Any())
-                    {
-                        foreach (var proc in allSteamProcs)
-                        {
-                            try { if (!proc.HasExited) proc.Kill(); } catch { }
-                        }
-                    }
-
-                    // Terminate existing Decky Loader instances (PluginLoader_noconsole)
-                    var deckyProcs = Process.GetProcessesByName("PluginLoader_noconsole");
-                    if (deckyProcs.Any())
-                    {
-                        foreach (var proc in deckyProcs)
-                        {
-                            try { if (!proc.HasExited) proc.Kill(); } catch { }
-                        }
-                    }
-
-                    // Give the OS a moment to free up file handles and ports
-                    await Task.Delay(1500);
-
-                    // Resolve the dynamic path to the user's homebrew folder
-                    string userProfileFolder = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-                    string deckyPath = Path.Combine(userProfileFolder, "homebrew", "services", "PluginLoader_noconsole.exe");
-
-                    if (File.Exists(deckyPath))
-                    {
-                        Debug.WriteLine("[GCM] Launching PluginLoader...");
-                        Process.Start(new ProcessStartInfo
-                        {
-                            FileName = deckyPath,
-                            UseShellExecute = true,
-                            WindowStyle = ProcessWindowStyle.Hidden // Keep the background clean
-                        });
-
-                        // Wait for the plugin loader to initialize its hooks before firing up Steam
-                        await Task.Delay(2000);
-                    }
-                    else
-                    {
-                        Debug.WriteLine($"[GCM] WARNING: Decky Loader executable not found at {deckyPath}");
-                    }
-
-                    // Since we forcibly closed Steam above, we MUST guarantee a cold boot for Steam now
-                    forceRestart = true;
-                    isSteamRunning = false;
-                }
-                // --- END DECKY LOADER INTEGRATION ---
-
-                // --- CASE A: COLD START (Boot or Steam was closed) ---
-                if (forceRestart || !isSteamRunning)
-                {
-                    Debug.WriteLine("[GCM] Steam Cold Boot into Big Picture Mode...");
-
-                    // If Decky was OFF, but we still need a force restart, ensure Steam is dead
-                    if (!useDeckyLoader && forceRestart)
-                    {
-                        var steamProcs = Process.GetProcessesByName("steam")
-                                                .Concat(Process.GetProcessesByName("steamwebhelper"))
-                                                .ToList();
-                        if (steamProcs.Any())
-                        {
-                            foreach (var proc in steamProcs) { try { if (!proc.HasExited) proc.Kill(); } catch { } }
-                            await Task.Delay(1500);
-                        }
-                    }
-
-                    RenameSteamStartupVideo_Start();
-                    StartProcessAsNonAdmin(steamExePath, "-gamepadui");
-                }
-                // --- CASE B: WARM START (Switching via Launcher Card) ---
-                else
-                {
-                    Debug.WriteLine("[GCM] Steam is already running. Triggering Big Picture switch (Warmstart)...");
-                    Process.Start(new ProcessStartInfo("steam://open/gamepadui") { UseShellExecute = true });
-                    steamLiefSchon = true;
-                }
-
-                // --- WAIT FOR THE STEAM WINDOW ---
-                IntPtr steamHwnd = IntPtr.Zero;
-                int attempts = 0;
-                // Faster timeout if it's just a warm switch
-                int maxAttempts = steamLiefSchon ? 20 : 60;
-
-                while (attempts < maxAttempts)
-                {
-                    steamHwnd = FindSteamBigPictureWindow();
-                    if (steamHwnd != IntPtr.Zero)
-                    {
-                        await Task.Delay(800); // Let the UI build up
-                        break;
-                    }
-                    await Task.Delay(250);
-                    attempts++;
-                }
-
-                // --- MAXIMIZE AND FORCE TO FOREGROUND ---
-                if (steamHwnd != IntPtr.Zero)
-                {
-                    Debug.WriteLine($"[GCM] Steam BP window ready. Applying Nuclear-Focus...");
-                    await ForcefullyBringToForeground(steamHwnd);
-                    ShowWindow(steamHwnd, 3); // SW_SHOWMAXIMIZED
-                }
-                else
-                {
-                    Debug.WriteLine("[GCM] Timeout! Steam BP window was not found.");
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[GCM] Error in StartSteam: {ex.Message}");
-            }
-        }
-
-        private async Task StartPlaynite()
-        {
-            try
-            {
-                // 1. Läuft die Fullscreen-App vielleicht schon? Dann nur nach vorne holen!
-                Process proc = Process.GetProcessesByName("Playnite.FullscreenApp").FirstOrDefault();
-                if (proc != null && proc.MainWindowHandle != IntPtr.Zero)
-                {
-                    Debug.WriteLine("[GCM] Playnite Fullscreen läuft bereits. Bringe in den Vordergrund...");
-                    MakeSelfNonTopmost();
-                    await ForcefullyBringToForeground(proc.MainWindowHandle);
-                    return;
-                }
-
-                // 2. WICHTIG: Die Desktop-App zwingend beenden! 
-                // Sonst weigert sich Playnite oft, in den Fullscreen-Modus zu wechseln.
-                var desktopProcs = Process.GetProcessesByName("Playnite.DesktopApp");
-                foreach (var dp in desktopProcs)
-                {
-                    try
-                    {
-                        dp.Kill();
-                        await Task.Delay(200); // Kurz warten, bis sie wirklich zu ist
-                    }
-                    catch { }
-                }
-
-                // 3. VOLLAUTOMATISCHE ERKENNUNG (Nur FullscreenApp)
-                string playnitePath = AutoDetectLauncherPath("playnite");
-
-                if (string.IsNullOrWhiteSpace(playnitePath) || !File.Exists(playnitePath))
-                {
-                    throw new FileNotFoundException("Playnite Fullscreen.exe konnte nicht gefunden werden.");
-                }
-
-                Debug.WriteLine($"[GCM] Starte Playnite Fullscreen von: {playnitePath}");
-                MakeSelfNonTopmost();
-
-                // 4. Starten mit WorkingDirectory und strikten Parametern
-                StartProcessAsNonAdmin(playnitePath, "--startfullscreen --hidesplashscreen", Path.GetDirectoryName(playnitePath));
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[GCM] Error in StartPlaynite: {ex.Message}");
-                await messagebox("Playnite Fullscreen could not be started. Please check the installation.");
-            }
-        }
-
-        private async Task StartGfn()
-        {
-            const int SW_SHOWMAXIMIZED = 3;
-            try
-            {
-                Debug.WriteLine("[GCM] Prüfe GeForce Now...");
-
-                // 1. Wenn GFN schon läuft
-                Process[] runningProcs = Process.GetProcessesByName("GeForceNOW");
-                foreach (var p in runningProcs)
-                {
-                    if (p.MainWindowHandle != IntPtr.Zero && IsWindowVisible(p.MainWindowHandle))
-                    {
-                        MakeSelfNonTopmost();
-                        await ForcefullyBringToForeground(p.MainWindowHandle);
-                        ShowWindow(p.MainWindowHandle, SW_SHOWMAXIMIZED);
-                        return;
-                    }
-                }
-
-                // 2. VOLLAUTOMATISCHE ERKENNUNG (Keine Settings mehr!)
-                string gfnPath = AutoDetectLauncherPath("gfn");
-
-                if (string.IsNullOrWhiteSpace(gfnPath))
-                {
-                    throw new FileNotFoundException("GeForce Now could not be detected automatically on this system.");
-                }
-
-                Process.Start(new ProcessStartInfo(gfnPath) { UseShellExecute = true });
-
-                // 3. Warten, bis das Fenster da ist
-                int attempts = 0;
-                IntPtr gfnHwnd = IntPtr.Zero;
-
-                while (attempts < 40)
-                {
-                    await Task.Delay(500);
-                    Process[] procs = Process.GetProcessesByName("GeForceNOW");
-                    foreach (var p in procs)
-                    {
-                        if (p.MainWindowHandle != IntPtr.Zero && IsWindowVisible(p.MainWindowHandle))
-                        {
-                            gfnHwnd = p.MainWindowHandle;
-                            break;
-                        }
-                    }
-                    if (gfnHwnd != IntPtr.Zero) break;
-                    attempts++;
-                }
-
-                if (gfnHwnd != IntPtr.Zero)
-                {
-                    MakeSelfNonTopmost();
-                    await ForcefullyBringToForeground(gfnHwnd);
-                    await Task.Delay(100);
-                    ShowWindow(gfnHwnd, SW_SHOWMAXIMIZED);
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[GCM] Fehler in StartGfn: {ex.Message}");
-                await messagebox("GeForce Now konnte nicht gefunden oder gestartet werden.");
-            }
-        }
-
-
-        private async Task StartOtherLauncher()
-        {
-            try
-            {
-                string launcherPath = AppSettings.Load<string>("customlauncherpath");
-
-                if (string.IsNullOrWhiteSpace(launcherPath) || !File.Exists(launcherPath))
-                {
-                    throw new FileNotFoundException("Der Pfad für den Custom Launcher ist ungültig oder wurde nicht gefunden.");
-                }
-
-                ShellManagementService.KillAllProcessesForExecutableFileName(Path.GetFileName(launcherPath));
-                StartProcessAsNonAdmin(launcherPath, null, Path.GetDirectoryName(launcherPath));
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"Fehler in StartOtherLauncher: {ex.Message}");
-                await messagebox("Der Custom Launcher konnte nicht gestartet werden. Bitte den Pfad in den Einstellungen prüfen.");
-                BackToWindows();
-            }
-        }
-
-
-
-
-        #endregion launcher
-        #region start
         // ERSETZE DEINE KOMPLETTE Start()-METHODE MIT DIESER VERSION
 
         #region start
@@ -5929,28 +5369,27 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
             switch (launcher)
             {
                 case "steam":
-                    await StartSteam(true); // <--- HIER: true für Kaltstart beim Boot!
+                    await _launcherService.StartSteamAsync(true).ConfigureAwait(false);
                     break;
                 case "playnite":
-                    await StartPlaynite();
+                    await _launcherService.StartPlayniteAsync().ConfigureAwait(false);
                     break;
                 case "custom":
-                    await StartOtherLauncher();
+                    await _launcherService.StartOtherLauncherAsync().ConfigureAwait(false);
                     break;
                 case "gfn":
-                    await StartGfn();
+                    await _launcherService.StartGfnAsync().ConfigureAwait(false);
                     break;
                 case "xbox":
                     await StartXbox(true);
                     break;
                 default:
                     AppSettings.Save("launcher", "steam");
-                    await StartSteam(true); // <--- HIER AUCH: true
+                    await _launcherService.StartSteamAsync(true).ConfigureAwait(false);
                     break;
             }
         }
 
-        #endregion start
         #endregion start
         #region TaskManager
         /// <summary>
@@ -11326,6 +10765,7 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
         private static readonly IntPtr HWND_TOP = IntPtr.Zero;
         private const uint SWP_NOSIZE = 0x0001;
         private const uint SWP_NOMOVE = 0x0002;
+        private const uint SWP_NOACTIVATE = 0x0010;
 
         [DllImport("user32.dll")]
         private static extern bool ShowWindowbtf(IntPtr hWnd, int nCmdShow);
@@ -11594,48 +11034,17 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
             {
                 try { RootGrid?.Focus(FocusState.Programmatic); } catch { /* focus not critical */ }
             });
+
+            DeferredInitialization.RunAfterFirstPaint(DispatcherQueue, () =>
+            {
+                Debug.WriteLine("[DeferredInit] Post–first paint.");
+                GitHubUpdateCheck.RunOptionalFireAndForget();
+            });
         }
 
         // --- STEAM INJECTION ---
 
-        public static void RenameSteamStartupVideo_Start()
-        {
-            try
-            {
-                if (!IsGcmVideoEnabled()) return;
-                if (!IsSteamInjectionEnabled()) return;
-
-                // Auto-Detect statt Settings!
-                string steamPath = AutoDetectLauncherPath("steam");
-                if (string.IsNullOrEmpty(steamPath)) return;
-
-                string moviesPath = Path.Combine(Path.GetDirectoryName(steamPath), "steamui", "movies");
-                if (!Directory.Exists(moviesPath)) return;
-
-                string myVideo = Path.Combine(moviesPath, "GCM_vid.webm");
-                string steamOriginal = Path.Combine(moviesPath, "bigpicture_startup.webm");
-                string steamBackup = Path.Combine(moviesPath, "bigpicture_startup.old.webm");
-
-                if (File.Exists(steamBackup))
-                {
-                    if (File.Exists(steamOriginal)) File.Delete(steamOriginal);
-                    File.Move(steamBackup, steamOriginal);
-                }
-
-                if (!File.Exists(myVideo)) return;
-
-                if (File.Exists(steamOriginal))
-                {
-                    File.Move(steamOriginal, steamBackup);
-                }
-                File.Copy(myVideo, steamOriginal, true);
-                Debug.WriteLine("[SteamInjection] Swap erfolgreich.");
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[SteamInjection] Start-Fehler: {ex.Message}");
-            }
-        }
+        public static void RenameSteamStartupVideo_Start() => LauncherService.RenameSteamStartupVideo_Start();
 
         public static void RenameSteamStartupVideo_End()
         {
@@ -11821,7 +11230,7 @@ private static readonly string SettingsFilePath = Path.Combine(SettingsFolder, "
                 }
                 else
                 {
-                    StartProcessAsNonAdmin(app.FilePath);
+                    LauncherService.StartProcessAsNonAdmin(app.FilePath);
                 }
 
                 // Clean up the UI after launching.
