+++ wingamepad/README.md
@@ -1,0 +1,6 @@
+# wingamepad
+
+Legacy **WPF** utility that polls XInput and fires global hotkeys. It is **not** part of the core DeckTop shell (`gcmloader`).
+
+- Prefer integrating gamepad behavior in **gcmloader** when `use_controller_navigation` is enabled.
+- This project may be removed after parity review; see repo `docs/PRIORITY.md` / product decision before deleting.
